requirements.txt
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
google-api-core==1.16.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
lxml==4.4.2
mysql-connector-python==8.0.18
ntlm-auth==1.4.0
numpy==1.18.1
pandas==0.25.3
protobuf==3.11.2
psycopg2==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rsa==4.0
SharePlum==0.2.0
six==1.13.0
smbus2==0.3.0
spidev==3.4
urllib3==1.25.7
